pyproject.toml
[project]
name = "identity-service"
version = "0.1.0"
description = "SSO Identity Service with FastAPI"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "python-multipart==0.0.6",
    "sqlalchemy==2.0.25",
    "alembic==1.13.1",
    "asyncpg>=0.30.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-dotenv==1.0.0",
    "pydantic[email]>=2.9.0",
    "pydantic-settings>=2.5.0",
    "httpx==0.26.0",
    "authlib==1.3.0",
    "fastapi-mail==1.4.1",
    "python-dateutil==2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "black==23.12.1",
    "ruff==0.1.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
